requirements.txt
asgiref==3.7.2
contourpy==1.2.0
# customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
Django==4.2.7
# django-admin-datta==1.0.10
django-api-generator==1.0.17
django-datetime-widget==0.9.3
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
djangorestframework==3.14.0
# fonttools==4.46.0
# kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psycopg2==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
